Se crean Test Junit para la entidad Puntos de Recolección. Se crea el atributo coordenada en Punto de Recolección y se elimina Zona

+++ com.sgr/src/main/java/com/sgr/api/VehiculoControllerw.java
@@ -29,7 +29,7 @@ public class VehiculoControllerw {
 	}
 
 	@GetMapping("/vehiculo/{id}")
-	public Vehiculo getUsuario(@PathVariable int id) {
+	public Vehiculo getUsuario(@PathVariable Long id) {
 		return vehiculoServiceImplement.getById(id);
 	}
